requirements.txt
pandas==2.0.3
geopandas==0.14.0
shapely==2.0.2
psycopg2-binary==2.9.9
openpyxl==3.1.2
python-dotenv==1.0.0